cmd-tool/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
idna==3.3
python-decouple==3.6
requests==2.28.1
typer==0.6.1
urllib3==1.26.10
